Move SubmissionAnswer to components folder
