+++ client/app/bundles/course/assessment/submission/containers/SubmissionEditForm.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { FieldArray, reduxForm } from 'redux-form';
 
-import SubmissionAnswer from './SubmissionAnswer';
+import SubmissionAnswer from '../components/SubmissionAnswer';
 
 class SubmissionEditForm extends Component {
